requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.8.0
torch>=1.10.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Optimization
optuna>=2.10.0
bayesian-optimization>=1.2.0

# Data Collection
yfinance>=0.1.70

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
pytest>=6.2.5
black>=21.12b0
flake8>=3.9.0

# Additional dependencies
newsapi-python==0.2.7
nltk==3.8.1
seaborn==0.13.0
python-dotenv==1.0.0
ta==0.10.2
statsmodels==0.14.0
prophet==1.1.4
scipy==1.11.4
plotly==5.18.0
dash==2.14.1
dash-bootstrap-components==1.5.0
pytest==7.4.3
mypy==1.7.1
tweepy==4.14.0
textblob==0.17.1
praw==7.7.1
hyperopt==0.2.7
transformers==4.35.2 